Estimated parameters.

This function estimates parameters like mean, standard deviation for a gamlss object. This function is very useful when the mean or standard deviation do not match with mu, sigma or some function of mu and/or sigma.

estimated_parameter(mod, fun = "mean", m = 100)

Arguments

mod
a gamlss object.

fun
the estimated parameter, by default is mean but can be change for sd or var.

m
a number to indicate the number of observations to simulate, by default its values is 100.

Value

estimated_parameter function returns a vector.

Details

The function obtains the fitted values for mu, sigma, nu and tau. The functions simulates m observations and then it obtains the mean/fun defined by the user.

Examples

# Example for NOrmal regression ------------------------
n <- 50
x <- runif(n=n)
y <- rnorm(n=n, mean=-5 + 10 * x, sd=exp(1 + 2 * x))
library(gamlss)

#> Loading required package: splines

#> Loading required package: gamlss.data

#> 
#> Attaching package: 'gamlss.data'

#> The following object is masked from 'package:datasets':
#> 
#>     sleep

#> Loading required package: gamlss.dist

#> Loading required package: MASS

#> Loading required package: nlme

#> Loading required package: parallel

#>  **********   GAMLSS Version 5.1-4  ********** 

#> For more on GAMLSS look at http://www.gamlss.org/

#> Type gamlssNews() to see new features/changes/bug fixes.

mod <- gamlss(y ~ x, sigma.fo=~x, family=NO)

#> GAMLSS-RS iteration 1: Global Deviance = 340.4941 
#> GAMLSS-RS iteration 2: Global Deviance = 339.6621 
#> GAMLSS-RS iteration 3: Global Deviance = 339.661 
#> GAMLSS-RS iteration 4: Global Deviance = 339.661 

# To explore the coefficients
coef(mod, "mu")

#> (Intercept)           x 
#>   -6.737415   16.015761 

coef(mod, "sigma")

#> (Intercept)           x 
#>    1.108047    1.776101 

# Obtaining the means
m1 <- fitted(mod, what="mu")
m2 <- estimated_parameter(mod, m=10000, fun="mean")
# Obtaining the standard deviations
s1 <- fitted(mod, what="sigma")
s2 <- estimated_parameter(mod, m=10000, fun="sd")

# Comparing the estimated means
cor(m1, m2)

#> [1] 0.9999088

# Comparing the estimated standard deviations
cor(s1, s2)

#> [1] 0.9998705

# Example for GAmma regression --------------------------
n <- 50
x <- runif(n=n)
y <- rGA(n=n, mu=exp(-5 + 10 * x), sigma=exp(-1 + 2 * x))
mod <- gamlss(y ~ x, sigma.fo=~x, family=GA)

#> GAMLSS-RS iteration 1: Global Deviance = -9.4469 
#> GAMLSS-RS iteration 2: Global Deviance = -17.5708 
#> GAMLSS-RS iteration 3: Global Deviance = -17.6536 
#> GAMLSS-RS iteration 4: Global Deviance = -17.6538 

# To explore the coefficients
coef(mod, "mu")

#> (Intercept)           x 
#>   -5.095443    9.634598 

coef(mod, "sigma")

#> (Intercept)           x 
#>   -1.026522    2.093336 

# Obtaining the means
m1 <- fitted(mod, what="mu")
m2 <- estimated_parameter(mod, m=10000, fun="mean")
# Obtaining the standard deviations
s1 <- fitted(mod, what="sigma")
s2 <- estimated_parameter(mod, m=10000, fun="sd")

# Comparing the estimated means
cor(m1, m2)

#> [1] 0.9996386

# Comparing the estimated standard deviations
# in GA parameterization standard devition = mu * sigma
cor(s1*m1, s2)

#> [1] 0.9996582