Estimation of mean and/or variance for models fitted with gamlss.

This function estimates parameters like mean, standard deviation for a gamlss object. This function is very useful when the mean or standard deviation do not match with mu, sigma or some function of mu and/or sigma.

est_param(mod, fun = "mean", m = 100)

Arguments

mod

a gamlss object.

fun

the estimated parameter, by default is mean but can be change for sd or var.

m

a number to indicate the number of observations to simulate, by default its values is 100.

Value

est_param function returns a vector.

Details

The function obtains the fitted values for mu, sigma, nu and tau. The functions simulates m observations and then it obtains the mean/fun defined by the user.

Examples

# Example for NOrmal regression ------------------------
# Here E(y) = mu and Var(y)= sigma
n <- 7
x <- runif(n=n)
y <- rnorm(n=n, mean=-5 + 10 * x, sd=exp(1 + 2 * x))
library(gamlss)
#> Loading required package: splines
#> Loading required package: gamlss.data
#> 
#> Attaching package: 'gamlss.data'
#> The following object is masked from 'package:datasets':
#> 
#>     sleep
#> Loading required package: gamlss.dist
#> Loading required package: nlme
#> Loading required package: parallel
#>  **********   GAMLSS Version 5.4-20  ********** 
#> For more on GAMLSS look at https://www.gamlss.com/
#> Type gamlssNews() to see new features/changes/bug fixes.
mod <- gamlss(y ~ x, sigma.fo=~x, family=NO)
#> GAMLSS-RS iteration 1: Global Deviance = 43.4364 
#> GAMLSS-RS iteration 2: Global Deviance = 35.2613 
#> GAMLSS-RS iteration 3: Global Deviance = 34.1276 
#> GAMLSS-RS iteration 4: Global Deviance = 34.0948 
#> GAMLSS-RS iteration 5: Global Deviance = 34.0943 

# Obtaining the means
m1 <- fitted(mod, what="mu")
m2 <- est_param(mod, m=10000, fun="mean")
cbind(m1, m2) # Comparing the estimated means
#>            m1         m2
#> 1 -6.09961525 -6.0990137
#> 2  5.96700504  9.0397923
#> 3  2.22696999  2.6019221
#> 4 -4.87533943 -4.8755479
#> 5 -7.27413102 -7.2749168
#> 6  0.07543462  0.1363465
#> 7  0.57796394  0.4918223
cor(m1, m2)
#> [1] 0.9899216

# Obtaining the standard deviations
s1 <- fitted(mod, what="sigma")
s2 <- est_param(mod, m=10000, fun="sd")
cbind(s1, s2) # Comparing the estimated standard deviations
#>             s1           s2
#> 1   0.13810096   0.13986106
#> 2 276.18534671 274.26679377
#> 3  26.18534313  26.31902698
#> 4   0.29861774   0.29864206
#> 5   0.06590076   0.06603477
#> 6   6.75264134   6.75616393
#> 7   9.26719010   9.29111948
cor(s1, s2)
#> [1] 0.9999994

# Example for GAmma regression --------------------------
# Here E(y) = mu and Var(y)= sigma^2 mu^2
n <- 7
x <- runif(n=n)
y <- rGA(n=n, mu=exp(-5 + 10 * x), sigma=exp(-1 + 2 * x))
mod <- gamlss(y ~ x, sigma.fo=~x, family=GA)
#> GAMLSS-RS iteration 1: Global Deviance = 16.9264 
#> GAMLSS-RS iteration 2: Global Deviance = 16.451 
#> GAMLSS-RS iteration 3: Global Deviance = 16.4508 

# Obtaining the means
m1 <- fitted(mod, what="mu")
m2 <- est_param(mod, m=10000, fun="mean")
cbind(m1, m2) # Comparing the estimated means
#>             m1           m2
#> 1 424.15608788 436.31392501
#> 2   2.06654083   2.04519647
#> 3   0.01677800   0.01668257
#> 4 283.71710580 292.39557310
#> 5   0.15756978   0.15690060
#> 6   0.05551859   0.05555330
#> 7 804.44551589 759.73614636
cor(m1, m2)
#> [1] 0.998866

# Obtaining the standard deviations
s1 <- fitted(mod, what="sigma") * fitted(mod, what="mu")
s2 <- est_param(mod, m=10000, fun="sd")
cbind(s1, s2) # Comparing the estimated standard deviations
#>             s1           s2
#> 1 1.034661e+03 1.035823e+03
#> 2 2.104148e+00 2.150535e+00
#> 3 7.753993e-03 7.816998e-03
#> 4 6.478880e+02 6.206000e+02
#> 5 1.051675e-01 1.049267e-01
#> 6 3.122518e-02 3.139663e-02
#> 7 2.179634e+03 2.101418e+03
cor(s1, s2)
#> [1] 0.9998398